+++ WebCollector/src/main/java/cn/edu/hfut/dmic/webcollector/util/Config.java
@@ -26,17 +26,17 @@ public class Config {
 
     public static String DEFAULT_USER_AGENT = "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:36.0) Gecko/20100101 Firefox/36.0";
     public static int MAX_RECEIVE_SIZE = 1000 * 1000;
-    public static long THREAD_KILLER = 1000 * 60 * 2;
-    public static long WAIT_THREAD_END_TIME = 1000 * 60;
+    public static int THREAD_KILLER = 1000 * 60 * 2;
+    public static int WAIT_THREAD_END_TIME = 1000 * 60;
     /*最大连续重定向次数*/
     public static int MAX_REDIRECT = 2;
 
-    public static long TIMEOUT_CONNECT = 3000;
-    public static long TIMEOUT_READ = 10000;
+    public static int TIMEOUT_CONNECT = 3000;
+    public static int TIMEOUT_READ = 10000;
     public static int MAX_EXECUTE_COUNT = 10;
 //    public static String DEFAULT_HTTP_METHOD = "GET";
     public static int TOP_N = 0;
 
-    public static long EXECUTE_INTERVAL = 0;
+    public static int EXECUTE_INTERVAL = 0;
 
 }
